-VBO and FBO externalization Now TilePrimitive class doesn't care about hardware. All is done via OpenGLStuff

+++ trunk/src/zildo/fwk/filter/FadeFilter.java
@@ -11,7 +11,7 @@ public class FadeFilter extends ScreenFilter {
 	}
 	
 	public void preFilter() {
-		float coeff=1.0f - ((float) getFadeLevel() / 256.0f);
+		float coeff=1.0f - (getFadeLevel() / 256.0f);
 		GL11.glColor4f(coeff, coeff, coeff, 1.0f);
 	}
 }
